avoid canonicalization unless needed
if debcargo encounters a Cargo.toml file in a .crate archive that pre-dates the
automatic canonicalization, it attempts to canonicalize it itself. this happens
before patches are applied, but after extraction including handling of file
exclusions.

since it is possible that a file-exclusion requires a corresponding patch to
Cargo.toml for the latter to be valid (which is a requirement for
canonicalization), move the canonicalization call into an if-branch that at
least limits it being called only when needed, instead of always but discarding
the result most of the time.

Fixes: #44

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>