stop issuing a now-useless warning about collapse_features being enabled
When collapse_features is enabled, debcargo currently issues a long bold
warning saying it should not be used.  However, in reality it is exactly
the opposite, there should be a very strong reason to NOT collapse
features.  This has been recommended to do for quite some time, and more,
today, ftpmasters (rightfully) rejects packages without features collapsed.
These days, debcargo should warn in the opposite case, when collapsing
features is NOT enabled.

Just remove the now-useless warning.

Signed-off-by: Michael Tokarev <mjt@tls.msk.ru>