canonicalization: adapt to cargo 0.82 changes
cargo now requires readme and license_file to point at existing files when canonicalizing Cargo.toml. this is not the case for a lot of ancient crate versions, so debcargo has to work around this by adapting or removing such references if they are invalid and manual canonicalization is required.

Signed-off-by: Fabian Grünbichler <git@fabian.gruenbichler.email>